pyproject.toml
[project]
name = "python_x509_pkcs11"
description = "Python library for signing x509 using keys in an pkcs11 device such as a HSM"
readme = "README.md"
requires-python = ">=3.6"
version = "0.1.83"
keywords = ["x509", "pkcs11", "HSM"]
license = {file = "LICENSE"}
authors = [
    { name = "Victor Näslund", email = "victor@sunet.se" }
]

dependencies = [
    "asn1crypto",
    "python-pkcs11"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
exclude = [
  "/.github",
  "/docs",
  "*.mypy_cache*",
  "*__pycache__*",
  "*.csr",
  "*.crl",
  "*.crt",
  "*.key",
  ".#*",
  "*~"
]

[tool.hatch.build.targets.wheel]
packages = ["src/python_x509_pkcs11"]

[project.urls]
Homepage = "https://github.com/SUNET/python_x509_pkcs11"

